pyproject.toml
[tool.poetry]
name = "trumo-task"
version = "0.1.0"
description = ""
authors = ["Tim Koval <timkoval00@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^3.0.2"
python-dotenv = "^1.0.1"
PyJWT = "^2.8.0"
Flask-Login = "^0.6.3"
pydantic = "^2.6.3"
pymongo = "^4.6.2"
pydantic-mongo = "^2.1.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
